fix: Cells over critical mass losing some mass
Previously, if a cell had 5 or more mass while its critical mass was 4,
4 of its mass would be distributed among its neighboring cells as usual.
However, the 1 or more leftover mass would disappear, as the original
cell would unconditionally be turned into an empty cell, thus losing
mass in the process.

This has been fixed so that leftover mass will remain in the original
cell.